requirements.txt
gitlab
click
GitPython
pandas
matplotlib
seaborn
tabulate
python-gitlab